Update 03-1ERS-CONTACTS-AVEC-BACKBONE.md
The french verb "adresser" has not the same meaning as the english verb "to address". Sorry to bother with such a detail but I ear that mistake too often around me, it has became kind of a private war.

Tremedous reading otherwise, I'm learning a lot thanks to your work.